website_app/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
passlib[bcrypt]
psycopg2-binary
jinja2
starlette 
PyJWT
itsdangerous
